pyproject.toml
[tool.poetry]
name = "documentation-generator"
version = "0.1.0"
description = ""
authors = ["nikhil.pandey"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
GitPython = "^3.1.43"
Pygments = "^2.18.0"
boto3 = "^1.34.148"
tree-sitter = "^0.22.3"
javalang = "^0.13.0"
esprima = "^4.0.1"
clang = "^17.0.6"
openai = "^1.37.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
